Limpiar código muerto y unificar la base de tiempo del contrato
Resultado de una revisión de calidad de todo vision-system/. Sin cambios
de comportamiento: se borra lo que no se usaba y se unifica una
duplicación real.

Borrados verificados como no usados (sin tests, sin __all__, y ningún
despacho dinámico del proyecto puede alcanzarlos):

- rectificador_desde_config() en geometry/distorsion.py, que además
  estaba rota: llamaba a Calibracion.ruta_perfil, un método que no
  existe desde que los perfiles pasaron a ser por cámara. Nadie la
  llamó nunca, por eso el AttributeError jamás se disparó.
- Compatibilidad.hay_problema y SistemaCoordenadas.celdas_a_mm.
- Cuatro imports sin leer: sys, dataclasses.field, perfiles_disponibles
  y BLANCO (estos dos en las dos ramas de su try/except).
- Dos variables muertas: id_paginas en patron_calibracion, que ahora se
  usa en vez de repetir el literal "2 0 R"; y los vectores de
  _DIRECCIONES en precision_ubicacion, que no hacían falta porque la
  medición compara la distancia |B - A| y no el sentido.

Duplicación factorizada:

- ahora_ms() pasa a schema.py. El simulador y el cliente de referencia
  tenían cada uno su copia, y son la misma base de tiempo del protocolo:
  quien escribe ts_ms y quien mide la latencia contra él tienen que
  coincidir siempre. Es aditivo —no toca ningún campo, nombre, unidad ni
  semántica del JSON—, así que NO es un cambio de contrato y no sube
  PROTOCOL_VERSION.

Se conservó a propósito lo que parece muerto y no lo es: celda_a_pixel
(documentada en sources/README.md), el Protocol FuenteImagen (descrito
en tres READMEs) y los __enter__/__exit__ de las tres fuentes, que son
el arreglo pendiente de las fugas de cámara. Se dejaron separadas las
duplicaciones que son conceptos distintos, como el ahora_ms de
sources/fuente.py, que es el instante de captura de un Cuadro y no el
sello del protocolo.

Verificado tras los borrados: geometría sintética TODO OK, los PDF del
patrón salen byte por byte idénticos, simulador y cliente extremo a
extremo sin mensajes inválidos ni saltos, y el contrato completo
corriendo bajo el Python 3.9.6 de fábrica de macOS, que es justamente
el caso que protege el piso bajo.
