+++ vision-system/vision/tools/panel.py
@@ -28,7 +28,7 @@ from __future__ import annotations
 
 import sys
 import unicodedata
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 
 import cv2
 import numpy as np
